+++ react/components/SearchResult.js
@@ -28,6 +28,7 @@ class SearchResult extends Component {
     // on SSR the getDerivedStateFromProps isn't called
     products: this.props.products,
     recordsFiltered: this.props.recordsFiltered,
+    facetRecordsFiltered: this.props.facetRecordsFiltered,
     brands: this.props.brands,
     map: this.props.map,
     params: this.props.params,
@@ -63,6 +64,7 @@ class SearchResult extends Component {
       const {
         products,
         recordsFiltered,
+        facetRecordsFiltered,
         brands,
         map,
         params,
@@ -76,6 +78,7 @@ class SearchResult extends Component {
       return {
         products,
         recordsFiltered,
+        facetRecordsFiltered,
         brands,
         map,
         params,
@@ -122,6 +125,7 @@ class SearchResult extends Component {
     const {
       mobileLayoutMode,
       recordsFiltered,
+      facetRecordsFiltered,
       products = [],
       brands,
       map,
@@ -137,7 +141,11 @@ class SearchResult extends Component {
     const term =
       params && params.term ? decodeURIComponent(params.term) : undefined
 
-    if (recordsFiltered === 0 && !loading) {
+    // only show the not found page if the reason for it
+    // isn't because of some applied filter, but that we
+    // *really* don't have any products to show for the
+    // current query
+    if (facetRecordsFiltered === 0 && !loading) {
       return <ExtensionPoint id="not-found" term={term} />
     }
 
@@ -153,7 +161,7 @@ class SearchResult extends Component {
               <ExtensionPoint id="breadcrumb" {...breadcrumbsProps} />
             </div>
           )}
-          <ExtensionPoint id="search-title" params={params} />
+          <ExtensionPoint id="search-title" params={params} map={map} products={products} />
           <ExtensionPoint
             id="total-products"
             recordsFiltered={recordsFiltered}